pyproject.toml
[tool.poetry]
name = "backend-services-trucks"
version = "1.0.0"
description = ""
authors = ["Cesar Juarez <cesar@wordbox.ai>"]
packages = [
    { include = "src" }
]


[tool.poetry.dependencies]
python = "^3.8"
google-cloud-logging = "^1.15.1"
neo4j = "4.1.1"
fastapi = "^0.61.1"
uvicorn = "^0.11.8"
python-dotenv = "^0.14.0"

[tool.poetry.dev-dependencies]
pytest = "^6.0.1"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
